Makes new TT04 implementation available.
From email with Matthias, it is necessary to pass the diffusion and
external stimulus current to the (TT04) reaction model.  The signature
of the calc method in the Reaction class has been modified appropriately
and the implementation of the BlueBeats version of TT04
(IBM_tenTusscher04) has been restored to actually use the external
current.  The simulation loop passes the currents to the reaction model.



git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/trunk@60 baecf98e-1e81-4516-8920-d4c889e67ea4