This version passes spot checks for comparison of JHU conductivities and diffusion weights as compared to BlueBeats.
The method of initializing the block of tissue cells and block of
conductivities in the Salheen98Diffusion code was fixed to more closely
resmeble what is done in BlueBeats.

Conductivity models now must supply a default tensor that can be
calculated with no arguments.  This is used in Salheen98Diffusion to
initialize the conductivity block.



git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/trunk@115 baecf98e-1e81-4516-8920-d4c889e67ea4