Add nThreads template files to the test directory.
This isn't very well organized, but it does provide a way to look at the
code performance as threads are added.  I don't expect this will be
useful without modifications, but it is a starting point and I needed a
place to stash it that wasn't on a temporary filesystem.



git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@470 baecf98e-1e81-4516-8920-d4c889e67ea4