Splits FGRDiffusion into three classes.  The OMP and Threads versions are unoptimized and should be stable for production use.  The FGRDiffusion class contains the SIMD code that we are actively working on optimizing.  There is still a substantial amount of duplicated code in the various classes.  I'll work on combining it at a later date.
The code will automatically select a variant of FGR depending on the
value of parallelDiffusionReaction.

This solves the problem that Diffusion hangs on BGQ.

I've tested (lightly) on Sierra, BGP, and BGQ and the production stuff
seems to be working ok.  The SIMD version on BGQ is not giving the
right answer.



git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@515 baecf98e-1e81-4516-8920-d4c889e67ea4