By Tomas Oppelstrup, Apr 30, 2012: I have modified simiulationLoop.cc to use one barrier per squad in the barrier between the calls to the non-gate and the gate kernels.
I have also added a new header file, fastBarrier_nosync.hh. It is
included by simulationLoop.cc.

The per squad barrier is enabled by the inclusion of the new file
fastBarrier_nosync.hh in simulationLoop.hh. Removing the #include of
fastBarrier_nosync.hh is identical to running the old barrier code.

In the case fast barriers are enabled (e.g. the fastBarrier.hh has
been included, and the corresponding macro is set by that file),
per-squad barriers will be used. Otherwise the old code is used.

If the fastBarrierBGQ.hh file is included (as is done by
fastBarrier.hh on the BGQ architecture), then the memory
syncronization instruction is omitted, since it should not be
necessary for a synchronization within one core (on BGQ one squad
consists of the four threads of one core -- this is assumed).

This has been tested by comparing the data file generating by running
the old code on the blockperformance spi_244_grid_64 case for 100k
timesteps of size 1us, to that generated by running the new code on
the same test case. The new code was run 100 times, and generated
data files identical to that of the old code each time.

Both in the old and the new code, there is a timer around this
barrier, the GateNonGate timer. The new code reduces the time reported
by this timer by about 0.8us.




git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@662 baecf98e-1e81-4516-8920-d4c889e67ea4