Removed the non-BGQ implementation of the nosync fast barriers and defined PER_SQUAD_BARRIER only on the BGQ build.
This keeps in place the current squad level barriers on BGQ, but reverts
to the usual non-optimzed barriers on other platforms.  More thought
needs to go into the barrier implementations.  In particular, the
current BGQ implementation is not fully correct.  It won't handle the
case where different cores have different squad sizes.  This is not a
common case, and it may never come up, but we should either handle it
properly or make sure it doesn't happen.



git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@734 baecf98e-1e81-4516-8920-d4c889e67ea4