Fix problem introduced in r959 that turning on certain sensors would cause a core dump.
The problem was caused by the fact that MPI is not thread safe.  Before
r959 loopIO and the halo exchange were both called by thread 0 on the
diffusion threads.  Hence there was no chance that both would be active
at the same time.  With loopIO now called on the reaction threads we can
have a situation where the halo exchange is active at the same time as
sensors and other I/O related code that might be using MPI.

For now the fix is to add a critical section around the body of the
loopIO function and around the start and wait functions of the halo
exchange.  Out of an abundance of caution I have put critical sections
around the SPI halo exchange as well.  There is little doubt this will
hurt performance.

This problem is another data point to keep in mind as we take the time
to redesign the IO code in the coming days.  A proper redesign should
allow us to remove the critical sections.



git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@979 baecf98e-1e81-4516-8920-d4c889e67ea4