Added code in Simulate.hh and TT06Dev_Reaction.cc to change the memory allocation sizes so that the problem of nodes with particular numbers of cells running slowly.
The new code is guarded by a macro defined in the new file slow_fix.hh.
As of checkin, the macro definition is commented out, so that the new
code is turned off. Enable by uncommenting the SLOW_FIX macro definition.

The new code, when turned on, is expected to improve nonGate-times for
nodes with unfavourable numbers of cells by 4us, assuming about 5300
cells per node.




git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@1100 baecf98e-1e81-4516-8920-d4c889e67ea4