Big reductions in the overhead associated with Stimulus.  The Stimulus time should have been completely covered by the Halo Exchange, but these improvements still have value as the trim several unneeded operations.
1.  Stimulus objects that stimulate no cells on a node are now dropped
instead of being added to the stimulus list.  This is a huge reduction
in function call overhead to call functions that never do anything.

2.  Zeroing array for stimulus is now threaded.

3.  Stimulus returns a value indicating whether the stimulus was
non-zero.  We only zero out the dVmD array on steps where the stim was
non-zero in the previous step.

With these changes, Stimulus is much cheaper on the timesteps when no
stimulus occurs and should be covered by the halo exchange on the <3%
of steps when on.



git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@1187 baecf98e-1e81-4516-8920-d4c889e67ea4