Introduced new precomputed work partition for the non gate code, so that it matches the same cells per core as in the gate code. This same new work partitioning is used for the integrator. This enables the use of per core barriers between gate and non gate code, and before the integrator. The total gain on a 16x16x14 blockPerformance run is about 0.54us.
With this update, the non gate code seem to run slower,
which is compensated for by a faster barrier times, and
a slightly faster integrator.

For the 48x48x14 blockPerformance, where the barriers
matter less, this code gives a loss of about 0.9us,
or just under 0.25%.

This new code can be turned off (including the per core barriers)
by defining the LEGACY_NG_WORKPARTITION macro in slow_fix.hh.



git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/EP/trunk@1188 baecf98e-1e81-4516-8920-d4c889e67ea4