More difficulties with the dynamic range.
For now, I've set the min range to 1e-7 of the max range.  This
avoids problems where the graph is really flat so nothing will
fit.  No idea if this will mess with exponentials where
we really need more than 1e-7 dynamic range. The rational behind this
change is that a single polynomial really doesn't work well
when you need the same coefficients to give more than 1e-7 accuracy
requirements for different parts of the function.